Refactor Logger from core to utils
